Refactor 'setFilter' used in RatingsFilter and SearchFilter with new hook useListFilter
