+++ frontend/src/features/ratings/RatingsFilter.tsx
@@ -1,34 +1,18 @@
 import React, { useState } from 'react';
-import { useLocation, useHistory } from 'react-router-dom';
-
 import { Collapse, Grid, Box } from '@material-ui/core';
 
 import { FiltersButton } from 'src/components';
+import { useListFilter } from 'src/hooks';
 import IsPublicFilter from './IsPublicFilter';
 import MarkAllRatingsMenu from './MarkAllRatings';
 
 function RatingsFilter() {
   const [expanded, setExpanded] = useState(false);
-  const Location = useLocation();
-  const history = useHistory();
-  const searchParams = new URLSearchParams(Location.search);
+  const [filterParams, setFilter] = useListFilter();
 
   const handleExpandClick = () => {
     setExpanded(!expanded);
   };
-
-  function setFilter(key: string, value: string) {
-    if (value) {
-      searchParams.set(key, value);
-    } else {
-      searchParams.delete(key);
-    }
-    // Reset pagination if filters change
-    if (key !== 'offset') {
-      searchParams.delete('offset');
-    }
-    history.push({ search: searchParams.toString() });
-  }
 
   return (
     <Box color="text.secondary">
@@ -39,7 +23,7 @@ function RatingsFilter() {
         <Grid container spacing={4} style={{ marginBottom: '8px' }}>
           <Grid item xs={12} sm={6} md={4}>
             <IsPublicFilter
-              value={searchParams.get('isPublic') ?? ''}
+              value={filterParams.get('isPublic') ?? ''}
               onChange={(value) => setFilter('isPublic', value)}
             />
           </Grid>
