+++ frontend/src/features/recommendation/SearchFilter.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
-import { useLocation, useHistory } from 'react-router-dom';
 import { Collapse, Grid, Box, makeStyles } from '@material-ui/core';
 
 import { FiltersButton } from 'src/components';
+import { useListFilter } from 'src/hooks';
 import LanguageFilter from './LanguageFilter';
 import DateFilter from './DateFilter';
 import CriteriaFilter from './CriteriaFilter';
@@ -16,26 +16,11 @@ const useStyles = makeStyles({
 function SearchFilter() {
   const classes = useStyles();
   const [expanded, setExpanded] = useState(false);
-  const Location = useLocation();
-  const history = useHistory();
-  const searchParams = new URLSearchParams(Location.search);
+  const [filterParams, setFilter] = useListFilter();
 
   const handleExpandClick = () => {
     setExpanded(!expanded);
   };
-
-  function setFilter(key: string, value: string) {
-    if (value) {
-      searchParams.set(key, value);
-    } else {
-      searchParams.delete(key);
-    }
-    // Reset pagination if filters change
-    if (key !== 'offset') {
-      searchParams.delete('offset');
-    }
-    history.push({ search: searchParams.toString() });
-  }
 
   return (
     <Box color="text.secondary">
@@ -44,13 +29,13 @@ function SearchFilter() {
         <Grid container spacing={4} className={classes.filtersContainer}>
           <Grid item xs={6} md={3} lg={2}>
             <DateFilter
-              value={searchParams.get('date') ?? ''}
+              value={filterParams.get('date') ?? ''}
               onChange={(value) => setFilter('date', value)}
             />
           </Grid>
           <Grid item xs={6} md={3} lg={2}>
             <LanguageFilter
-              value={searchParams.get('language') ?? ''}
+              value={filterParams.get('language') ?? ''}
               onChange={(value) => setFilter('language', value)}
             />
           </Grid>
