+++ frontend/src/hooks/index.ts
@@ -1,2 +1,3 @@
 export { useSearchParams } from './useSearchParams';
 export { useLoginState } from './useLoginState';
+export { useListFilter } from './useListFilter';
